# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: KACP API
  version: 0.1.0-draft
  description: |
    KCC Agent Cloud Platform 외부 API (0단계 초안).
    규칙과 내부 API는 docs/design/04-api.md 참고.
    상태 변경 요청은 X-KACP-CSRF 헤더가 필요하다.
servers:
  - url: https://app.kacp.cloud/api/v1
security:
  - session: []
...
paths:
...
  /users/search:
    get:
      tags: [org]
      summary: 사람 찾기 (멤버 추가용)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: departmentId, in: query, schema: { type: string, format: uuid } }
        - { name: includeDescendants, in: query, schema: { type: boolean, default: true } }
      responses:
        '200':
          description: 최대 20명
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/UserRef' }

  # ───────────── names ─────────────
...
  /admin/users:
    get:
      tags: [admin]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { $ref: '#/components/schemas/PlatformRole' } }
        - { name: status, in: query, schema: { type: string, enum: [active, disabled, must_change_password] } }
        - { name: team, in: query, schema: { type: string } }
        - { name: departmentId, in: query, schema: { type: string, format: uuid } }
        - { name: includeDescendants, in: query, schema: { type: boolean, default: true } }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/AdminUser' }
    post:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email: { type: string, format: email }
                name: { type: string }
                departmentId: { type: string, format: uuid }
                title: { type: string }
                employeeNo: { type: string }
                platformRole: { $ref: '#/components/schemas/PlatformRole' }
                initialPassword: { type: string, description: 생략 시 자동 생성 }
                teams:
                  type: array
                  items:
                    type: object
                    properties:
                      team: { type: string }
                      teamRole: { $ref: '#/components/schemas/TeamRole' }
      responses:
        '201':
          description: 생성됨. 초기 비밀번호는 이 응답에서만 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/AdminUser' }
                  initialPassword: { type: string }
        '409': { $ref: '#/components/responses/Error' }
  /admin/users/{userId}:
    parameters: [{ $ref: '#/components/parameters/UserId' }]
    get:
      tags: [admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AdminUser'
                  - properties:
                      recentLogins:
                        type: array
                        items:
                          type: object
                          properties:
                            at: { type: string, format: date-time }
                            ip: { type: string }
                            success: { type: boolean }
    patch:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                departmentId: { type: [string, 'null'], format: uuid }
                title: { type: [string, 'null'] }
                employeeNo: { type: [string, 'null'] }
                platformRole: { $ref: '#/components/schemas/PlatformRole' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
  /admin/users/{userId}/reset-password:
    parameters: [{ $ref: '#/components/parameters/UserId' }]
    post:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      responses:
        '200':
          description: 임시 비밀번호(1회 표시)
          content:
            application/json:
              schema:
                type: object
                properties:
                  temporaryPassword: { type: string }
  /admin/users/{userId}/{action}:
    parameters:
      - { $ref: '#/components/parameters/UserId' }
      - { name: action, in: path, required: true, schema: { type: string, enum: [disable, enable] } }
    post:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                restartTeams: { type: boolean, description: 열린 WebSocket을 끊기 위해 소속 팀 컨테이너 재시작 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
  /admin/users/bulk-department:
    post:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userIds, departmentId]
              properties:
                userIds: { type: array, items: { type: string, format: uuid } }
                departmentId: { type: string, format: uuid }
      responses:
        '204': { description: 변경됨 }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: kacp_session
  parameters:
    Csrf:
      name: X-KACP-CSRF
      in: header
      required: true
      schema: { type: string }
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
  responses:
    Error:
      description: 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [NAME_TAKEN] }
            message: { type: string, description: 사용자에게 보여줄 한국어 문구 }
            details: { type: object }
    Page:
      type: object
      properties:
        nextCursor: { type: [string, 'null'] }
...
    PlatformRole:
      type: string
      enum: [admin, user]
    TeamRole:
      type: string
      enum: [team_admin, member]
...
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        departmentName: { type: [string, 'null'], description: 동명이인 구분용 표시 }
    DepartmentRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        pathNames: { type: array, items: { type: string }, description: "최상위부터 (예: [경영지원본부, 인사팀])" }
...
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
        department: { oneOf: [{ $ref: '#/components/schemas/DepartmentRef' }, { type: 'null' }] }
        title: { type: [string, 'null'] }
        employeeNo: { type: [string, 'null'] }
        platformRole: { $ref: '#/components/schemas/PlatformRole' }
        status: { type: string, enum: [active, disabled] }
        mustChangePassword: { type: boolean }
        teams:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              teamRole: { $ref: '#/components/schemas/TeamRole' }
        lastLoginAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...